added basic controllers for course without any reference to other documents

+++ server/controller/category-controller.js
@@ -4,7 +4,8 @@ import { Category } from "../models/index.js";
 
 export const addCategories = async (req, res) => {
   try {
-    if (Category && Category.length) await Category.deleteMany({});
+    await Category.deleteMany({});
+    // Same process can be done using insertMany(categories)
     const newCategory = await Category.create(categories);
     return res.status(201).json(newCategory);
   } catch (error) {
@@ -29,6 +30,11 @@ export const getCategoriesById = async (req, res) => {
   try {
     const categoryById = req.params.id;
     const receivedCategory = await Category.findById(categoryById);
+    if (!receivedCategory) {
+      return res
+        .status(404)
+        .json({ message: `No category with ${categoryById} found` });
+    }
     return res.status(200).json(receivedCategory);
   } catch (error) {
     return res.status(404).json({
@@ -39,6 +45,12 @@ export const getCategoriesById = async (req, res) => {
 
 export const deleteAllCategories = async (req, res) => {
   try {
+    const categoryCount = Category.find();
+    if (!categoryCount.length) {
+      return res
+        .status(404)
+        .json({ message: `No categories available to delete` });
+    }
     await Category.deleteMany({});
     return res
       .status(201)
